Add Ubicloud IPv6 address space to the namespace
This commit adds the new address space "oo-bic-lood-safe-ce" to the tap
devices.  This address space will be used for providing private endpoint
to the VMs so that we can serve certificate and key files.

In relation to that, we are making another change to add some nftables
rules to the cloudinit.img. With the help of new nftables rules inside
the VM, we are blocking access to the "oo-bic-lood-safe" address space
if the user is not root which has skuid 0.