Add ParadeDB flavor to the PostgreSQL create page
As part of our partnership with ParadeDB, we're introducing ParadeDB as
a new flavor for our PostgreSQL database.

The commit adds ParadeDB logo, along with a partnership notice, to
enhance the visibility of our partnership to the create page.

The customer have to accept the partnership details before creating a
new database.