Support reserve cache without size information
It was assumed that reserve cache sends the size information, though
it is not valid for all cases (docker layer cache is an example). In that
case, reserve cache returns presigned urls for the biggest cache file size.
Size of the entry is updated during commit cache as it is always sent
with that request.