Update vm_host to support allocation of longer prefixes to VMs
We always tried to allocate /79 to the VM to be used as 2 /80s.This
doesn't work with the new leaseweb hosts because we already get a /80
from them for the host. Therefore, we need to be able to allocate much
smaller. Keeping the entropy the same (2 bytes), we change the ipv6
allocation logic. This way, hetzner hosts will continue allocation /79
while leaseweb allocates /95.