Use proxy ndp for both clover and guest ephemeral
With the recent change to use dnsmasq as the first level dns server, we
get out to the internet with the clover_ephemeral address. The existing
ndp proxy logic was only proxying the guest ephemeral which makes the VM
IPv6 stack working as expected. However, we realized that the dns
queries through dnsmasq was failing. Further investigation showed that
the dnsmasq was able to send packets out but no response was getting
back in. I realized the source address is the first half of the /95
prefix for dnsmasq. Therefore, we need to enable ndp_proxy for that
prefix, too. This way, our response packets started to flow back in.