Set correct HTTP headers during tests by default
Push down some currently-unused/leftover code that applies metadata to
tests, and then use it to set Content-Type and Accept headers for API
tests to the correct default.

Negative tests can override them a second time.

Currently, this doesn't break any tests, but when one enables checking
requests and responses to the openapi schema, it does.  This is a
backport from an development branch.